+++ tests/test_console_reporter.py
@@ -1,15 +1,19 @@
 import logbook
-
 import pytest
+
 from slash._compat import StringIO
 from slash.reporting.console_reporter import ConsoleReporter
 
 
-def test_console_reporter(suite, level, config_override):  # pylint: disable=redefined-outer-name
+@pytest.mark.parametrize('console_level', [True, False])  # pylint: disable=redefined-outer-name
+def test_console_reporter(suite, level, config_override, console_level):  # pylint: disable=redefined-outer-name
+    config_override('log.color_console', console_level)
     config_override('log.console_level', level)
     summary = suite.run()
     suite.add_test().when_run.raise_exception()
-    assert summary.get_console_output()
+    output = summary.get_console_output()
+    assert output
+    assert bool(console_level) == ('\x1b' in output)
 
 
 def test_silence_manual_errors(suite, suite_test, config_override):
@@ -19,10 +23,12 @@ def test_silence_manual_errors(suite, suite_test, config_override):
     output = suite.run().get_console_output()
     assert 'slash.add_error' not in output
 
+
 def test_num_collected_printed_once(suite, suite_test):  # pylint: disable=unused-argument
     assert len(suite) > 1
     output = suite.run().get_console_output()
     assert output.count('collected') == 1
+
 
 @pytest.mark.parametrize('long_headline', [True, False])
 @pytest.mark.parametrize('multiline', [True, False])